Moved project subscription utils to subscription file
