+++ tests/unit/test_projects/test_services.py
@@ -1,7 +1,7 @@
 from notifications.models import Notification
 from posts.models import Post
 from projects.permissions import ObjectPermission
-from projects.services.common import notify_project_subscriptions_question_open
+from projects.services.subscriptions import notify_project_subscriptions_question_open
 from questions.models import Question
 from tests.unit.test_questions.factories import create_question
 from tests.unit.test_posts.factories import factory_post
